assimp-3.1.1 build ok on vs2010

+++ common/mesh.h
@@ -67,13 +67,13 @@ public:
       string number;
       string name = textures[i].type;
       if (name == "texture_diffuse")
-        number = std::to_string(diffuseNr++);
+        number = std::to_string((long double)diffuseNr++);
       else if (name == "texture_specular")
-        number = std::to_string(specularNr++); // transfer unsigned int to string
+        number = std::to_string((long double)specularNr++); // transfer unsigned int to string
       else if (name == "texture_normal")
-        number = std::to_string(normalNr++); // transfer unsigned int to string
+        number = std::to_string((long double)normalNr++); // transfer unsigned int to string
       else if (name == "texture_height")
-        number = std::to_string(heightNr++); // transfer unsigned int to string
+        number = std::to_string((long double)heightNr++); // transfer unsigned int to string
 
       // now set the sampler to the correct texture unit
       glUniform1i(glGetUniformLocation(shader.ID, (name + number).c_str()), i);
